requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
filelock==3.8.2
future==0.18.2
huggingface-hub==0.11.1
idna==3.4
more-itertools==9.0.0
numpy==1.23.5
packaging==22.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
whisper @ git+https://github.com/openai/whisper.git@0b5dcfdef7ec04250b76e13f1630e32b0935ce76
